pyproject.toml
[project]
name = "openai-realtime-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "keyboard>=0.13.5",
    "numpy>=2.2.4",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.0",
    "sounddevice>=0.5.1",
    "soundfile>=0.13.1",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
]
